@@ -38,9 +38,10 @@
   "devDependencies": {
     "chai": "^4.3.3",
     "cross-env": "^7.0.3",
-    "eslint": "^7.32.0",
+    "eslint": "^8.16.0",
     "eslint-config-digitalbazaar": "^2.8.0",
-    "eslint-plugin-jsdoc": "^32.3.4",
+    "eslint-plugin-jsdoc": "^39.3.2",
+    "eslint-plugin-unicorn": "^42.0.0",
     "mocha": "^10.0.0",
     "nyc": "^15.1.0"
   },
