@@ -23,7 +23,7 @@ const allImplementations = implementations;
  * @param {Function<boolean>} options.filter - A function to
  * filter the map's entries on that returns true or false.
  *
- * @returns {object.<string, Map>} Returns an object with matching
+ * @returns {Object<string, Map>} Returns an object with matching
  *   and non-matching maps.
  */
 export const filterImplementations = ({
@@ -55,7 +55,7 @@ export const filterImplementations = ({
  * @param {string} [options.property='issuers'] - The property to search for on
  *   an implementation.
  *
- * @returns {object.<string, Map>} Returns an object with matching
+ * @returns {Object<string, Map>} Returns an object with matching
  *   and non-matching maps.
  */
 export const filterByTag = ({
